libbladeRF_test: Fixed i686-w64-mingw32 warnings
Addressed the following warnings though use of casts to larger stdint.h types
and inttypes.h format specifiers:

test_timestamps.c: In function ‘counter_data_is_valid’:
test_timestamps.c:99:21: warning: unknown conversion type character ‘l’ in format [-Wformat=]
                     (unsigned long long) i, ctr, val);
                     ^
test_timestamps.c:99:21: warning: format ‘%x’ expects argument of type ‘unsigned int’, but argument 3 has type ‘long long unsigned int’ [-Wformat=]
test_timestamps.c:99:21: warning: too many arguments for format [-Wformat-extra-args]

test_tx_onoff.c:124:9: warning: unknown conversion type character ‘l’ in format [-Wformat=]
         printf("Sending burst @ %llu\n", (unsigned long long) meta.timestamp);
         ^
test_tx_onoff.c:124:9: warning: too many arguments for format [-Wformat-extra-args]

+++ host/libraries/libbladeRF_test/test_timestamps/src/test_timestamps.c
@@ -94,9 +94,9 @@ bool counter_data_is_valid(int16_t *samples, size_t n_samples, uint32_t ctr)
     for (i = 0; i < 2 * n_samples; i += 2, ctr++) {
         const uint32_t val = extract_counter_val((uint8_t*) &samples[i]);
         if (val != ctr) {
-            fprintf(stderr, "Invalid counter value @ sample %llu. "
+            fprintf(stderr, "Invalid counter value @ sample %"PRIu64". "
                     "Expected 0x%"PRIx32", got 0x%"PRIx32"\n",
-                    (unsigned long long) i, ctr, val);
+                    (uint64_t) i, ctr, val);
 
             return false;
         }
